# spec/openapi/uai-v1.yaml
openapi: 3.1.0
info:
  title: Universal Agent Identity API
  version: "0.1.0"
  summary: Identity, attestation and governance for autonomous AI agents.
  description: |
    The UAI v1 REST surface.

    REST and JSON because a protocol meant to be implemented by everyone must be
    implementable with `curl`. gRPC is used internally where call volume justifies it.

    **Authentication.** Every state-changing call requires proof of possession: mutual TLS
    with a SPIFFE X.509-SVID *and* an RFC 9421 HTTP message signature. mTLS proves the
    runtime instance; the message signature proves the persistent identity. Either alone is
    insufficient, because a stolen SVID has no persistent key and a stolen persistent key has
    no attested runtime.

    **What this API does not claim.** A verified identity is not a safe agent, and revocation
    does not stop code from running outside the ecosystem. See
    `docs/protocol/01-overview.md`.
  license:
    name: Apache-2.0 (implementation) / CC-BY-4.0 (specification)
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: UAI security
    email: security@uai.world
servers:
  - url: https://api.uai.world/v1
    description: Production API
  - url: https://verify.uai.world/v1
    description: Public verification endpoint (unauthenticated, cached)
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - mutualTLS: []
    messageSignature: []
paths:
...
  /capability-requests:
    post:
      tags: [action]
      operationId: requestCapability
      summary: Request a capability the agent does not hold
      description: |
        Creates a PENDING request and nothing else. It does not grant, and it cannot be made
        to: approval is an act by the human owner, out of band. There is no route in this API
        that writes a capability grant, and the database refuses one signed by the agent
        itself (22.9, T-11/T-13).

        Requires an RFC 9421 signature tagged `UAI-v1:capability-request`. One open request per
        (agent, capability): re-asking in a loop until an owner approves out of fatigue is a
        social attack on a human, and it is refused with 409.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [capability, justification]
              properties:
                capability: { type: string }
                justification:
                  type: string
                  description: |
                    What it is for, in terms the owner can decide on. Required: without it they
                    would be approving a capability name.
                purpose: { type: string }
                resource: { type: string }
      responses:
        '202':
          description: Request recorded, pending the owner's decision
          content:
            application/json:
              schema:
                type: object
                required: [request_id, capability, state, granted, note]
                properties:
                  request_id: { type: string }
                  capability: { type: string }
                  state: { enum: [PENDING] }
                  granted:
                    type: boolean
                    enum: [false]
                    description: Always false. Present so a caller reading this sees the answer.
                  requested_at: { type: string, format: date-time }
                  expires_at: { type: string, format: date-time }
                  note: { type: string }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /agents/{id}/capability-requests:
    get:
      tags: [action]
      operationId: listCapabilityRequests
      summary: An agent's capability requests
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent: { type: string }
                  requests: { type: array, items: { type: object } }
...
components:
  securitySchemes:
    mutualTLS:
      type: mutualTLS
      description: SPIFFE X.509-SVID. Proves the runtime instance.
    messageSignature:
      type: apiKey
      in: header
      name: Signature
      description: |
        RFC 9421 HTTP message signature over @method, @target-uri, content-digest,
        uai-agent-id and uai-nonce, tagged with the UAI domain. Proves the persistent
        identity. `agent_id` in a JSON body proves nothing.
    webAuthn:
      type: apiKey
      in: header
      name: X-UAI-Assertion
      description: |
        WebAuthn assertion from a hardware authenticator with user verification required.
        For votes, the challenge is the digest of the vote statement itself.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Required on every state-changing call. A replay with the same key and body returns the
        original response; the same key with a different body returns 409.
      schema: { type: string, minLength: 8, maxLength: 255 }
  responses:
    Problem:
      description: RFC 9457 problem document
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: |
        RFC 9457. Carries `decision_id` and `policy_version` whenever a policy was involved, so
        a refusal is auditable and reproducible rather than an opaque denial.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string, examples: [UAI_IDENTITY_REVOKED, UAI_POP_REQUIRED, UAI_CHAIN_CONFLICT] }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        decision_id: { type: string }
        policy_version: { type: string }
        governance_proof: { type: string }
        remediation: { type: string }
        trace_id: { type: string }
...